Move builder functions to libbuild
